refactor(transform): isolate shaker executable index
